Update overview components
