+++ components/Portfolio/Trading/Overview/index.tsx
@@ -3,159 +3,18 @@ import Graph from './Graph';
 import PositionGraph from './PositionGraph';
 import Equity from './Equity';
 import { defaults } from '@libs/Tracer';
-import styled from 'styled-components';
-import Dropdown from 'antd/lib/dropdown';
-import { Button } from '@components/General';
-import { Menu, MenuItem } from '@components/General/Menu';
 import ConnectOverlay from '@components/Overlay/ConnectOverlay';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
 import PositionOverlay from '@components/Overlay/PositionOverlay';
-
-interface HRowProps {
-    background?: string;
-    border?: boolean;
-}
-const HeadingRow = styled.div<HRowProps>`
-    display: flex;
-    align-items: center;
-    padding: 0 16px;
-    height: 60px;
-    width: 100%;
-    background: ${(props) => (props.background ? (props.background as string) : 'transparent')};
-    border-bottom: ${(props) => (props.border ? '1px solid var(--table-lightborder)' : 'none')};
-`;
-
-interface HPanelProps {
-    background?: string;
-}
-const HPanel = styled.div<HPanelProps>`
-    position: relative;
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    height: auto;
-    width: 100%;
-    padding: 0 15px 16px;
-    background: ${(props) => (props.background ? (props.background as string) : 'transparent')};
-    .equityStats {
-        flex-basis: calc(60% - 8px);
-        &.show {
-            max-height: 420px;
-            tr:nth-child(2) td,
-            tr:nth-child(3) td,
-            tr:nth-child(4) td,
-            tr:nth-child(5) td {
-                opacity: 1;
-                border-color: var(--table-darkborder);
-            }
-        }
-    }
-
-    .pnlGraph {
-        flex-basis: calc(40% - 8px);
-        min-height: 342px;
-    }
-`;
-
-const Title = styled.h1`
-    font-size: var(--font-size-large);
-    letter-spacing: var(--letter-spacing-extra-small);
-    color: var(--color-text);
-    margin-right: 2rem;
-    padding: 0;
-    width: fit-content;
-    white-space: nowrap;
-    font-weight: inherit;
-`;
-
-const StyledTriangleDown = styled.img`
-    position: absolute;
-    right: 10px;
-    top: 50%;
-    transform: translate(0%, -50%);
-    height: 0.5rem;
-    transition: all 400ms ease-in-out;
-    display: inline;
-
-    &.rotate {
-        transform: rotate(180deg);
-        margin-top: -4px;
-    }
-`;
-
-const Counter = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-left: -10px;
-    border-radius: 20px;
-    background: #3da8f5;
-    color: #00156c;
-    width: 52px;
-    height: 32px;
-    font-size: var(--font-size-small);
-`;
-
-const VScrollContainer = styled.div`
-    height: auto;
-    overflow: hidden auto;
-`;
-
-const HScrollContainer = styled.div`
-    position: relative;
-    display: flex;
-    width: auto;
-    height: auto;
-    min-height: 30vh;
-    overflow: auto hidden;
-    padding: 16px 8px;
-    box-sizing: unset;
-`;
-
-type PDProps = {
-    setOptions: (val: number) => void;
-    option: number;
-    keyMap: Record<number, string>;
-    className?: string;
-};
-const PortfolioDropdown: React.FC<PDProps> = styled(({ className, setOptions, option, keyMap }: PDProps) => {
-    const [rotated, setRotated] = useState(false);
-    const menu = (
-        <Menu
-            onClick={({ key }: any) => {
-                setOptions(parseInt(key));
-                setRotated(false);
-            }}
-        >
-            {Object.keys(keyMap).map((key) => {
-                return (
-                    <MenuItem key={key}>
-                        <span>{keyMap[parseInt(key)]}</span>
-                    </MenuItem>
-                );
-            })}
-        </Menu>
-    );
-    const handleVisibleChange = (visible: boolean) => {
-        setRotated(visible);
-    };
-    return (
-        <Dropdown className={className} overlay={menu} placement="bottomCenter" onVisibleChange={handleVisibleChange}>
-            <Button height="medium">
-                {keyMap[option]}
-                <StyledTriangleDown className={rotated ? 'rotate' : ''} src="/img/general/triangle_down_cropped.svg" />
-            </Button>
-        </Dropdown>
-    );
-})`
-    position: relative;
-    padding-right: 8px;
-    margin: unset;
-    &:hover {
-        background: none;
-        color: var(--color-primary);
-    }
-`;
+import {
+    VScrollContainer,
+    PortfolioDropdown,
+    Title,
+    HScrollContainer,
+    Counter,
+    SectionHeader,
+    HPanel,
+} from '@components/Portfolio';
 
 interface OProps {
     fetchedTracers: any;
@@ -178,59 +37,57 @@ const Overview: FC<OProps> = ({ fetchedTracers }: OProps) => {
     };
 
     return (
-        <>
-            <VScrollContainer>
-                <HeadingRow background={'#00125D'}>
-                    <Title>Equity Breakdown</Title>
-                    <div className="flex justify-content-between">
-                        <PortfolioDropdown
-                            setOptions={setCurrentPortfolio}
-                            option={currentPortfolio}
-                            keyMap={portfolioKeyMap}
-                        />
-                        <PortfolioDropdown setOptions={setCurrentPNL} option={currentPNL} keyMap={pnlKeyMap} />
-                    </div>
-                </HeadingRow>
-                <HPanel background={'#00125D'}>
-                    <Equity
-                        className="equityStats"
-                        balances={fetchedTracers[0]?.getBalance() ?? defaults.balances}
-                        fairPrice={fetchedTracers[0]?.getFairPrice() ?? defaults.fairPrice}
-                        baseTicker={fetchedTracers[0]?.baseTicker ?? defaults.baseTicker}
-                        quoteTicker={fetchedTracers[0]?.quoteTicker ?? defaults.quoteTicker}
+        <VScrollContainer>
+            <SectionHeader background={`var(--color-background-secondary)`}>
+                <Title>Equity Breakdown</Title>
+                <div className="flex justify-content-between">
+                    <PortfolioDropdown
+                        setOptions={setCurrentPortfolio}
+                        option={currentPortfolio}
+                        keyMap={portfolioKeyMap}
                     />
-                    <Graph
-                        className="pnlGraph"
-                        title="Profit and Loss"
-                        background
-                        selectedTracerAddress={fetchedTracers[0]?.address ?? ''}
+                    <PortfolioDropdown setOptions={setCurrentPNL} option={currentPNL} keyMap={pnlKeyMap} />
+                </div>
+            </SectionHeader>
+            <HPanel background={`var(--color-background-secondary)`}>
+                <Equity
+                    className="equityStats"
+                    balances={fetchedTracers[0]?.getBalance() ?? defaults.balances}
+                    fairPrice={fetchedTracers[0]?.getFairPrice() ?? defaults.fairPrice}
+                    baseTicker={fetchedTracers[0]?.baseTicker ?? defaults.baseTicker}
+                    quoteTicker={fetchedTracers[0]?.quoteTicker ?? defaults.quoteTicker}
+                />
+                <Graph
+                    className="pnlGraph"
+                    title="Profit and Loss"
+                    background
+                    selectedTracerAddress={fetchedTracers[0]?.address ?? ''}
+                />
+                {!account ? <ConnectOverlay /> : null}
+            </HPanel>
+            <SectionHeader border={true}>
+                <Title>Withdrawals</Title>
+                <Counter>{fetchedTracers?.length}</Counter>
+            </SectionHeader>
+            <HScrollContainer>
+                {fetchedTracers.map((tracer: any, i: number) => (
+                    <PositionGraph
+                        key={`position-graph-${i}`}
+                        selectedTracerAddress={tracer?.address ?? ''}
+                        base={tracer?.getBalance().base ?? defaults.base}
+                        quote={tracer?.getBalance().quote ?? defaults.quote}
+                        market={tracer?.marketId}
+                        fairPrice={tracer?.getFairPrice() ?? defaults.fairPrice}
+                        maxLeverage={tracer?.getMaxLeverage() ?? defaults.maxLeverage}
                     />
-                    {!account ? <ConnectOverlay /> : null}
-                </HPanel>
-                <HeadingRow border={true}>
-                    <Title>Open Positions</Title>
-                    <Counter>{fetchedTracers?.length}</Counter>
-                </HeadingRow>
-                <HScrollContainer>
-                    {fetchedTracers.map((tracer: any, i: number) => (
-                        <PositionGraph
-                            key={`position-graph-${i}`}
-                            selectedTracerAddress={tracer?.address ?? ''}
-                            base={tracer?.getBalance().base ?? defaults.base}
-                            quote={tracer?.getBalance().quote ?? defaults.quote}
-                            market={tracer?.marketId}
-                            fairPrice={tracer?.getFairPrice() ?? defaults.fairPrice}
-                            maxLeverage={tracer?.getMaxLeverage() ?? defaults.maxLeverage}
-                        />
-                    ))}
-                    {!account ? (
-                        <ConnectOverlay />
-                    ) : fetchedTracers.length === 0 ? (
-                        <PositionOverlay tracers={fetchedTracers} showMarketPreview={true} />
-                    ) : null}
-                </HScrollContainer>
-            </VScrollContainer>
-        </>
+                ))}
+                {!account ? (
+                    <ConnectOverlay />
+                ) : fetchedTracers.length === 0 ? (
+                    <PositionOverlay tracers={fetchedTracers} showMarketPreview={true} />
+                ) : null}
+            </HScrollContainer>
+        </VScrollContainer>
     );
 };
